EZP-26057: extract PermissionService unit tests
